# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: 'Task Manager API'
  version: 1.0.0
paths:
...
  '/api/tasks/{id}':
    get:
      tags:
        - Tasks
      operationId: get_task_get
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '[0-9a-f]{8}-[0-9a-f]{4}-[13-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}'
      responses:
        '200':
          description: 'The task'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: 'No task has this id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
components:
  schemas:
    Problem:
      required:
        - type
        - title
        - status
        - detail
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        violations:
          type: array
          items:
            required:
              - propertyPath
              - title
            properties:
              propertyPath:
                type: string
              title:
                type: string
            type: object
      type: object
    Task:
      required:
        - id
        - title
        - description
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          description: 'Status name, for example "new"'
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      type: object
      additionalProperties: false
...